NewLanguageConstructs: split off NewOperators sniff
Split the `NewLanguageConstructs` sniff into two:
* `NewLanguageConstructs` addressing the namespace separator + ellipsis.
* `NewOperators` in the `Operators` category addressing pow, pow equals, spaceship, coalesce, coalesce equals.
