+++ PHPCompatibility/Sniffs/LanguageConstructs/NewLanguageConstructsSniff.php
@@ -37,97 +37,13 @@ class NewLanguageConstructsSniff extends AbstractNewFeatureSniff
             '5.3' => true,
             'description' => 'the \ operator (for namespaces)',
         ),
-        'T_POW' => array(
-            '5.5' => false,
-            '5.6' => true,
-            'description' => 'power operator (**)',
-        ), // Identified in PHP < 5.6 icw PHPCS < 2.4.0 as T_MULTIPLY + T_MULTIPLY.
-        'T_POW_EQUAL' => array(
-            '5.5' => false,
-            '5.6' => true,
-            'description' => 'power assignment operator (**=)',
-        ), // Identified in PHP < 5.6 icw PHPCS < 2.6.0 as T_MULTIPLY + T_MUL_EQUAL.
         'T_ELLIPSIS' => array(
             '5.5' => false,
             '5.6' => true,
             'description' => 'variadic functions using ...',
         ),
-        'T_SPACESHIP' => array(
-            '5.6' => false,
-            '7.0' => true,
-            'description' => 'spaceship operator (<=>)',
-        ), // Identified in PHP < 7.0 icw PHPCS < 2.5.1 as T_IS_SMALLER_OR_EQUAL + T_GREATER_THAN.
-        'T_COALESCE' => array(
-            '5.6' => false,
-            '7.0' => true,
-            'description' => 'null coalescing operator (??)',
-        ), // Identified in PHP < 7.0 icw PHPCS < 2.6.2 as T_INLINE_THEN + T_INLINE_THEN.
-        /*
-         * Was slated for 7.2, but still not implemented. PHPCS however does already tokenize it.
-         * @link https://wiki.php.net/rfc/null_coalesce_equal_operator
-         */
-        'T_COALESCE_EQUAL' => array(
-            '7.2' => false,
-            '7.3' => true,
-            'description' => 'null coalesce equal operator (??=)',
-        ), // Identified in PHP < 7.0 icw PHPCS < 2.6.2 as T_INLINE_THEN + T_INLINE_THEN + T_EQUAL and between PHPCS 2.6.2 and PHPCS 2.8.1 as T_COALESCE + T_EQUAL.
     );
 
-
-    /**
-     * A list of new language constructs which are not recognized in older PHPCS versions.
-     *
-     * The array lists an alternative token to listen for.
-     *
-     * @var array(string => int)
-     */
-    protected $newConstructsPHPCSCompat = array(
-        'T_POW'            => T_MULTIPLY,
-        'T_POW_EQUAL'      => T_MUL_EQUAL,
-        'T_SPACESHIP'      => T_GREATER_THAN,
-        'T_COALESCE'       => T_INLINE_THEN,
-        'T_COALESCE_EQUAL' => T_EQUAL,
-    );
-
-    /**
-     * Token translation table for older PHPCS versions.
-     *
-     * The 'before' index lists the token which would have to be directly before the
-     * token found for it to be one of the new language constructs.
-     * The 'real_token' index indicates which language construct was found in that case.
-     *
-     * If the token combination has multi-layer complexity, such as is the case
-     * with T_COALESCE(_EQUAL), a 'callback' index is added instead pointing to a
-     * separate function which can determine whether this is the targetted token across
-     * PHP and PHPCS versions.
-     *
-     * {@internal 'before' was chosen rather than 'after' as that allowed for a 1-on-1
-     * translation list with the current tokens.}}
-     *
-     * @var array(string => array(string => string))
-     */
-    protected $PHPCSCompatTranslate = array(
-        'T_MULTIPLY' => array(
-            'before'     => 'T_MULTIPLY',
-            'real_token' => 'T_POW',
-        ),
-        'T_MUL_EQUAL' => array(
-            'before'     => 'T_MULTIPLY',
-            'real_token' => 'T_POW_EQUAL',
-        ),
-        'T_GREATER_THAN' => array(
-            'before'     => 'T_IS_SMALLER_OR_EQUAL',
-            'real_token' => 'T_SPACESHIP',
-        ),
-        'T_INLINE_THEN' => array(
-            'callback'   => 'isTCoalesce',
-            'real_token' => 'T_COALESCE',
-        ),
-        'T_EQUAL' => array(
-            'callback'   => 'isTCoalesceEqual',
-            'real_token' => 'T_COALESCE_EQUAL',
-        ),
-    );
 
     /**
      * Returns an array of tokens this test wants to listen for.
@@ -138,11 +54,7 @@ class NewLanguageConstructsSniff extends AbstractNewFeatureSniff
     {
         $tokens = array();
         foreach ($this->newConstructs as $token => $versions) {
-            if (defined($token)) {
-                $tokens[] = constant($token);
-            } elseif (isset($this->newConstructsPHPCSCompat[$token])) {
-                $tokens[] = $this->newConstructsPHPCSCompat[$token];
-            }
+            $tokens[] = constant($token);
         }
         return $tokens;
     }//end register()
@@ -161,29 +73,6 @@ class NewLanguageConstructsSniff extends AbstractNewFeatureSniff
     {
         $tokens    = $phpcsFile->getTokens();
         $tokenType = $tokens[$stackPtr]['type'];
-
-        // Translate older PHPCS token combis for new constructs to the actual construct.
-        if (isset($this->newConstructs[$tokenType]) === false) {
-            if (isset($this->PHPCSCompatTranslate[$tokenType])
-                && ((isset($this->PHPCSCompatTranslate[$tokenType]['before'], $tokens[$stackPtr - 1]) === true
-                    && $tokens[$stackPtr - 1]['type'] === $this->PHPCSCompatTranslate[$tokenType]['before'])
-                || (isset($this->PHPCSCompatTranslate[$tokenType]['callback']) === true
-                    && call_user_func(array($this, $this->PHPCSCompatTranslate[$tokenType]['callback']), $tokens, $stackPtr) === true))
-            ) {
-                $tokenType = $this->PHPCSCompatTranslate[$tokenType]['real_token'];
-            }
-        } elseif ($tokenType === 'T_COALESCE') {
-            // Make sure that T_COALESCE is not confused with T_COALESCE_EQUAL.
-            if (isset($tokens[($stackPtr + 1)]) !== false && $tokens[($stackPtr + 1)]['code'] === T_EQUAL) {
-                // Ignore as will be dealt with via the T_EQUAL token.
-                return;
-            }
-        }
-
-        // If the translation did not yield one of the tokens we are looking for, bow out.
-        if (isset($this->newConstructs[$tokenType]) === false) {
-            return;
-        }
 
         $itemInfo = array(
             'name' => $tokenType,
@@ -248,59 +137,6 @@ class NewLanguageConstructsSniff extends AbstractNewFeatureSniff
     {
         $data[0] = $errorInfo['description'];
         return $data;
-    }
-
-
-    /**
-     * Callback function to determine whether a T_EQUAL token is really a T_COALESCE_EQUAL token.
-     *
-     * @param array $tokens   The token stack.
-     * @param int   $stackPtr The current position in the token stack.
-     *
-     * @return bool
-     */
-    private function isTCoalesceEqual($tokens, $stackPtr)
-    {
-        if ($tokens[$stackPtr]['code'] !== T_EQUAL || isset($tokens[($stackPtr - 1)]) === false) {
-            // Function called for wrong token or token has no predecesor.
-            return false;
-        }
-
-        if ($tokens[($stackPtr - 1)]['type'] === 'T_COALESCE') {
-            return true;
-        }
-        if ($tokens[($stackPtr - 1)]['type'] === 'T_INLINE_THEN'
-            && (isset($tokens[($stackPtr - 2)]) && $tokens[($stackPtr - 2)]['type'] === 'T_INLINE_THEN')
-        ) {
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Callback function to determine whether a T_INLINE_THEN token is really a T_COALESCE token.
-     *
-     * @param array $tokens   The token stack.
-     * @param int   $stackPtr The current position in the token stack.
-     *
-     * @return bool
-     */
-    private function isTCoalesce($tokens, $stackPtr)
-    {
-        if ($tokens[$stackPtr]['code'] !== T_INLINE_THEN || isset($tokens[($stackPtr - 1)]) === false) {
-            // Function called for wrong token or token has no predecesor.
-            return false;
-        }
-
-        if ($tokens[($stackPtr - 1)]['code'] === T_INLINE_THEN) {
-            // Make sure not to confuse it with the T_COALESCE_EQUAL token.
-            if (isset($tokens[($stackPtr + 1)]) === false || $tokens[($stackPtr + 1)]['code'] !== T_EQUAL) {
-                return true;
-            }
-        }
-
-        return false;
     }
 
 }//end class
