+++ PHPCompatibility/Tests/LanguageConstructs/NewLanguageConstructsUnitTest.inc
@@ -2,13 +2,4 @@
 
 $foobar = new \Foobar();
 
-$x = 5 ** 10;
-$x **= 10;
-
-echo $x ?? $a;
-
-$x ??= 5;
-
-echo 1 <=> 1;
-
 function newTest(...$arguments) {}
