+++ PHPCompatibility/Tests/Operators/NewOperatorsUnitTest.inc
@@ -1,0 +1,9 @@
+<?php
+
+$x = 5 ** 10;
+$x **= 10;
+
+echo $x ?? $a;
+$x ??= 5;
+
+echo 1 <=> 1;
